Feat: Add section tabs and PDF export to Lost Time page
- Refactors the `LostTimeTrackingPage` by creating a new `LostTimeEntries` component to encapsulate the logic for displaying, filtering, and exporting lost time entries.
- Implements section-wise tabs to filter entries by section.
- Adds a date range filter with a "Search" button to trigger filtering.
- Updates the PDF export functionality to generate a report with data grouped by section into separate tables.
